Implemented search sort functionality

+++ controllers/user/deleteUser.js
@@ -10,10 +10,10 @@ module.exports = async function (userId, dbName) {
     const User = getTenantModel(dbName, 'User', userSchema);
     const InviteMember = getTenantModel(dbName, 'InviteMember', inviteMemberSchema);
 
-    const deletedUser = await User.findOneAndDelete({ _id: userId });
+    const deletedUser = await User.findOneAndDelete({ _id: userId, role: { $ne: 'Admin' } });
+    if (!deletedUser) throw new createHttpError(STATUS_CODE.NOT_FOUND, ERROR_MESSAGE.USER_NOT_FOUND);
     await TenantUserMap.deleteOne({ email: deletedUser.email });
     await InviteMember.deleteOne({ email: deletedUser.email });
-    if (!deletedUser) throw new createHttpError(STATUS_CODE.NOT_FOUND, ERROR_MESSAGE.USER_NOT_FOUND);
     await redis.del(`user:${dbName}`);
     return deletedUser;
 }
